Update Flow to v0.243.0
This errors on deprecated stuff by default which means
a lot of things that were deprecated ages ago are
now erroring.

A few of the %checks were entirely useless. The others
are replaced by type guards. It was impossible (AFAICT)
to replace the isDateEmpty one directly, so this adds
an isDateNonEmpty function and uses it where we were
previously negating isDateEmpty - this mirrors our existing
empty and notEmpty functions.

Similarly useless was the $Call for N_ln. We know the type
so we can just use it directly.

Uses of $ObjMap were replaced with mapped types or just removed
when useless, such as in the type for an unused redux function.

https://github.com/facebook/flow/releases/tag/v0.243.0