MBS-13740: Split report strings into reports translation component
Reports are a lot less critical (and include a lot longer strings) than
most of the rest of the Server translation component.

This first commit makes the necessary code changes
to move the translations to a new reports component.
In next commits I'll generate the relevant .pot file,
and generate start .po files to save the currently translated lines.