MBS-13697: Use right domain more in timeline page
Several parts of the timeline page were using the mb_server
translation domain, but clearly should be using the statistics one.
This change means the stats tabs and the hardcoded labels
for the legend both should use the right translations now.

Since the 'macros' files were just being used for a couple
tiny things now in the two TT files we have left for stats,
I just moved the calls inside them to the appropriate places.