Update eslint-plugin-import to 2.31.0
This is supposed to work with eslint9 - we'll see.

It allows a new checkTypeImports option for import/extensions
which we might as well include.

Some new warnings were being shown for /unambiguous and /no-cycle.
/unambiguous turns out to be way too whiny and we decided to
turn it off. /no-cycle required a fairly large rewriting of the
guess case code, which was a mess and is certainly somewhat better
(if not quite yet *good*) because of it.