Restore now working no-named-as-default rules
These were broken with flat config but now work again.

no-named-as-default-member did hit a lot of issues, mostly with the way
we access DBDefs properties. There's no real reason why we need
a default export for DBDefs files, really, and it seems cleaner anyway
to just import the one or two properties we need in most cases
and do import * as DBDefs for the few cases where we use three
or more properties.

As such, this also autogenerates DBDefs-client-browser so that
it contains the same exports as DBDefs-client.

There seems to be no real reason why unformatTrackLength
and calculateDiscID were exported separately while nothing else
in utils was - this just makes them consistent with everything
else in the file.