Update to eslint 9.12.0
This introduces a crash in no-unused-vars in some flow type pages
(https://github.com/eslint/eslint/issues/18876)
which might or might not get fixed on eslint or hermes.
That said, for now we are fine just disabling the no-unused-vars rule
on the three affected files.

A few new issues were detected for no-unused-vars for errors.
I explicitly indicated now that we're ignoring the error by using the variable
name ignoredError and ignoring that name for the rule.

As a new feature, this version of eslint now lets us know where the rules that
we are disabling are no longer warning or erroring (probably due to now fixed
eslint bugs). As such, we can remove those disable lines.