MBS-9253: Use type child_order rather than id to sort RGs
This will allow us to list EP release groups above singles, which
has been requested forever, and it makes sense in all markets except
maybe Japan. Given Japan would want singles before albums,
which we don't do anyway, this is probably not too relevant.

For secondary types, we were sorting by type id on _fast sorts
and by type name on _slow. This changes both to use child_order.

We still want type ids for filtering and as a fallback when
child_order is the same, so this adds child_order in addition
to them rather than replacing them.