MBS-13784: Support paid streaming for beatport
Beatport now supports streaming (with a subscription) for what seems
to be most, but possibly not all, of its content (there's a streaming flag
in the API, suggesting some music is still available for download only).

This keeps always selecting purchase for download, but also allows
selecting streaming alongside it for releases and recordings.
For now artists and labels are left as purchase only; since adding a separate
relationship for Beatport is not an available solution at the moment,
this seems like the least bad option.