Use - instead of --- as a slug fallback
We have 58 --- URLs, but around 500 with just - (probably because of
Harmony). There seems to be no reason to prefer --- so let's do -
in the future.
This will become more relevant as I need to clean up dj.beatport releases
which have the same release ID but do not have a slug.