MBS-13784: Clean up dj.beatport links
These are for a DJ-oriented UI for the beatport store, allowing live DJ
use. The content seems to be the same as the main store though and
with both allowing stream and purchase, there seems to be
no reason not to consolidate them.

Artists and labels have a slug on both versions of the site, so they can
be converted trivially. Releases don't, so we need the false slug
hack we already used for some older redirects. Track links are release id
plus track id, but I chose to clean them to tracks rather than releases
so that there is a way for users to get track links from the DJ site.
Hopefully it won't be too confusing since the UI mostly highlights
the release even from DJ track links.