MBS-13770: Allow admins to auto-approve and auto-reject any edit
There's plenty of cases where a removal of a spam release for example
sits open for days, and closing it earlier requires asking for more eyes on it
which does seem like the opposite of what you want to do with spam.

This allows account admins to use the same accept/reject mechanism
we already used in test servers to immediately accept or reject
any open edit (including their own).

To make what has happened clear, a special vote ("Admin approval" or
"Admin rejection") is added to the edit when closing it.
"Admin approval" is treated as an approval for icons and headers,
while "Admin rejection" is treated as a No vote.
The new vote types are searchable in edit searches, their stats can be seen
on admin profiles, etc.