MBS-13781: Support browsing genres by collection in the API
We added the feature in the last schema change but I completely forgot
the WS side of it. This implements it, plus some tests.

I had to move genre_all to be the last one defined so that Catalyst
would choose it first and avoid breaking genre lists, apparently;
see https://metacpan.org/dist/Catalyst-Runtime/view/lib/Catalyst/RouteMatching.pod