MBS-13564: Correctly override primary on alias locale change
Attempting to change the locale of an alias marked as primary to one
which already had a primary alias crashed with an ISE.

I noticed the issue is that we were not clearing the previous primary
value first for this case, since we were only looking for a changed
primary_for_locale. In this case
the locale changes, not the primary flag.

This ensures primary is cleared upon either primary or locale changes
if primary is being set or if it was already set and was not overriden
by the row changes.