Bump chrome-for-testing to 130.0.6723.91
Recent Selenium builds have been failing with

  WebDriverError: disconnected: not connected to DevTools
    (failed to check if window was closed: disconnected: not connected to DevTools)

even though we haven't changed anything that I'm aware of. Trying to update our
Chrome dependencies here to see if that helps in any way.

Also bump selenium-webdriver to 4.26.0; I've combined it with this commit
because the changelog suggests it's somewhat tied to specific Chrome versions:
https://github.com/SeleniumHQ/selenium/blob/trunk/javascript/node/selenium-webdriver/CHANGES.md