Remove flaky autocomplete test
This started failing during CI runs recently, and I have no clue why. (If you
run the tests manually in your browser, it will also fail if you click on the
page before it executes.)

I'm just removing the test because the old autocomplete code is on the way out
anyway.