Sort special icons before per-site when editing links
On the sidebar, we only show the default per-site icon if none of the
special icons apply. On the editing page, we were doing the other way
around. Consolidating it seems sensible and if so we should do
the special icons first, since overriding per-page ones is part
of their whole point.