Move getFaviconClass out of render and simplify
While I'm not sure if this is faster than three nested .some calls,
it shouldn't be a lot slower and I'd say it makes the code at least somewhat
more clear.

This does no longer prioritize blog over review, but that probably doesn't
matter since I'm not sure those can exist at the same time for the same
entity type and even if they could, something being a blog and review
for the same entity at the same time would just be bad data, so it's irrelevant
which icon it shows.