Extract and generalize load_aliases
We want to be able to load aliases even in cases where there is no need
to load all the rest of the related info for an artist, such as in tag pages.
That means it makes sense to extract load_aliases as a separate
method that can be called by load_related_info.

While we only use this in artists right now, there's no big reason that it
needs to be that way. As such, this moves it out of Data::Artist into the
Alias role so it can be used by other entities as well.