MBS-13816: Update no-nag requirement to $5/mo
As part of the changes in the donation system it has been decided
to up the no-nag donation requirement to $5 a month (which
would also give access to user flairs in LB and I guess eventually maybe
in MB and elsewhere as well?).

This is calculated on the MetaBrainz site, but we do need to update the
string about it.