Drop unneeded passedHover
We never actually pass a hover, so all passedHover was doing here was
complicate things for no reason. I dropped it.

This makes it simpler to merge the two cases where we show disambiguation
on hover, which makes my brain hurt a bit less (marginally)
and will make it easier to show primaryAlias in a subsequent commit.