Split load_aliases from load_related_info and use in more places
There's no real reason why this needs to be part of load_related_info.
Splitting it allows us to just call it for all entities which support it on
Collection / Search / Series rather than having the call be repeated
under each relevant entity (also making it more trivial to expand
to other entities if we ever were to).