MBS-13416: Link to entity pages from LB buttons
Release links redirect to release groups, so the LB team advised to
just link to RGs directly.
Since the pages cannot be played if empty, we don't change the existing
behaviour where we hide the
buttons if no releases exist in an RG
or no tracks exist on a release.
We no longer care about the first release's MBID for RGs, just that it
exists, so this simplifies that parameter
to be boolean.

Also adding links for artists since those are now available. I did not
skip "empty" artists (with no releases nor recordings) because
we only have this data from the artist index but we want the buttons
to appear on all tabs, and it doesn't seem worth passing a boolean up
from the Perl layer just for this.

Recording pages do not yet exist, so we keep linking to the player
for recordings (and recording collections). Since recordings are
the only case where an array of MBIDs is actually supported,
I changed the code to just return nothing when passed an array
in any other case.