MBS-13839: Also sort by ID in cancel_edits_and_votes
This is supposed to happen in the opposite order as the edits were
entered but sorting by open_time alone does not consistently work -
edits 118415049 (Add series) and 118415050 (Add rel to the series)
have the exact same open time since they were opened as a set,
and the code tries to close the series first causing an ISE.
This only breaks for series AFAICT, since for other stuff we actually
remove the entity even if it has relationships.

Just sorting by ID would probably be enough, but just to play it extra safe,
let's do time first, then ID if time is the same.

Changed the spammer test to specifically test for this case.