MBS-13801: Remove limits on tag cloud subqueries (#3400)
These were introduced in ace41c641667b1c9d91872996f73cea959773f32
under the assumption exact numbers weren't too important
because we didn't show them, and the results were reasonable enough.

The first was already a wrong assumption at the time, since we
did show the counts on hover in the cloud page. It's even worse now,
since we have a list version with counts front and center.

The second assumption seems also a bit off on testing without the limits;
it brings a few entirely new tags to the list, since the current query
undercounts tags that are used a bit on many entity types
in comparison with tags used a lot on one entity type only.

The original worry why these limits were added (a slow query that got
a second or so faster with them) still applies, of course. That said, this
query runs once, and then is cached for 24h. It seems acceptable to have
one person a day wait a bit longer for their results in exchange for
actually correct results when the count differences are in the
thousands on the lower ranks of the lists (the top is of course used
everywhere and already has the right counts mostly).