MBS-13862: Unbreak rel foreign keys for entities of same type
When both entities had the same type, I was overwriting the keys
for entity 0 instead of adding the entity 1 ones alongside them.

This could be avoided, but since we don't actually need to be
careful with this anymore since I figured out how to use
load_subobjects to only try to look at ACs where they exist,
the easiest way to avoid the issue is to revert the changes
to the edit files.