MBS-13863: Compare alias with content, not entity name
In cases where we have a relationship credit, for example, it doesn't
make sense to repeat the alias if it matches the credit.
On the other hand, it actually *does* seem to make sense to show
the alias if it doesn't match the credit even if it does match the name
- an editor checking a Japanese or Russian release might appreciate
knowing who a Western artist shown under a local credit is.