Split Selenium tests into four jobs
These take a long time to complete, and that time will only increase as we add
more and more browser tests.

Add a way to run partitions/subsets of the tests in separate jobs. We use
greedy number partitioning [1] to partition the tests using their number of
commands as their size.

[1] https://en.wikipedia.org/wiki/Greedy_number_partitioning