MBS-13886: Don't crash on wrong link_type for paged rels
This was hit and logged on Sentry (leads to a TypeError on JS).
I assume someone was just playing around with the URL params, but
it seems to make sense to block this at the controller level and since
I cannot think of any other way of getting here than the user fiddling
400 seems like a good fit.

We were trying to do this at the JS level earlier, but it crashes
when trying to calculate the linkPhrase we were hoping to check
(possibly a new issue after some changes, or possibly it never
worked) so it never gets to return the error. It seems sensible
in any case to check on the Perl layer and just not bother
with any of the loading if there's a clear broken situation.