Reuse server strings when possible in stats
We were making people translate stuff twice a lot more than strictly
needed, and mwiencek suggested we don't. I'm sure I've missed some
reused strings, but this is most of them.

To ensure we don't accidentally use l for statistics strings,
this explicitly imports l as lMbServer and uses that.
We have a test to prevent bare l(), but it doesn't seem to work
at the moment since it wasn't triggered by this.