MBS-13870: Add filter for recordings not related to works
This allows filtering an artist's recording list by whether the recordings
have at least one work related to them, or they have none.

This is admittedly mostly useful for editing, but it can be *very* useful
for that, allowing users to easily find recordings that still need work
relationships - it would make it much easier to improve the discographies
of prolific classical performers, such as orchestras and conductors,
and probably be of use also for popular music.