MBS-13893: Block invisible characters (Hangul, Braille) in usernames
Invisible characters are problematic in usernames, since they make it
easy to pose as someone else. While there can be cases where these
characters could make sense elsewhere, in usernames we should
play it a bit safer. As such, this creates a separate sanitize_username
method that can call further cleanups on top of the basic sanitize,
and runs the pre-existing remove_invisible_characters in it.