MBS-13804: Also show where the review is on sidebar
For reviews, it can be annoying to have just a "Review" link with no info
about what kind of review it is and have to hover in order to find out.
This shows the hostname of the review URL on the sidebar, so that
the user can decide based on that what review(s) they are interested in.
I remove www. because it doesn't seem to add anything of value
and takes sidebar space.