MBS-13597: Add release filter to label index pages
This extends the release filter from the artist releases page to label index
pages, which have very similar lists. Most of the code can be reused,
so I moved some into roles.

The only real difference with the artist forms are that ACs are taken
for all the releases in the label and that "no label exists" is not a filter
option for labels, for obvious reasons.

Tests added, with very small changes from the artist release ones.