MBS-13770: Don't show admin approval buttons on edit lists (#3449)
From very quick personal experience, it's way too easy to misclick these
while trying to do a normal approve (or probably even when adding a note).
For the cases where this is needed, it's fine to spend the few extra seconds
to go into the edit page(s) and apply or reject from there.