MBS-8500: Add "not by me" filter for collection edit lists
Collection edit lists cannot be filtered, and it can be very hard
to find others' edits to review on a collection where you are fairly active
yourself as an editor.

This could ideally be eventually solved by edit searches, but we are
not really near to having collection filters for edit searches.

At least for now, having an extra "hide own edits" condition for collections
seems like by far the easiest choice, and combines well with the status
condition we already have in find_by_collection.