webpack: set `minChunks: 5` for common-chunks
https://webpack.js.org/plugins/split-chunks-plugin/#splitchunksminchunks

The common-chunks bundle includes too much code; we could allow more chunks to
be duplicated in other bundles to reduce the size of common-chunks, and thus
the amount of code that gets run on each page (at the cost of more files having
to be downloaded/cached initially).

5 is a bit of an arbitrary number, but is a modest step up from 2 and matches
what I chose for the vendors bundle.