Remove jed-data bundle
The jed-data bundle basically just contained an object which the other
language-specific bundles could import and extend. But it's not really needed,
because we can just put the language-specific data into `GLOBAL_JS_NAMESPACE`.

The other use of the jed-data module (which was the only file in the bundle)
was to provide an "en" template that other languages' data could be based on,
as well as serve as fallback data for the "en" case (since the `Jed`
constructor still needs something to initialize it with). I've just renamed the
file to jedDataTemplate.mjs and froze the exported object to make this
remaining use case clearer.