MBS-13918: Link to docs from primary alias checkbox label
Some users seem to just check primary for every alias they add,
whether it makes sense or not. It doesn't seem they understand
what the field is for. As such, this adds an easy to access link to the
specific guidelines for when to mark an alias as primary, in the hopes
that it will make users check first or at least it will be easier to point
out that they should read them.