Install xgettext-js from org-scoped package rather than git
Installing it from git is very likely to be the cause of issues such as
https://github.com/metabrainz/musicbrainz-docker/issues/295 which are outside
of our control, as we cannot guarantee that yarn will generate an identical
tarball on all platforms.

I've published our fork of xgettext-js to the metabrainz org on npm, and have
updated our code to reference that instead.

The only change I made to the fork in the meantime was bumping the version of
hermes-parser used.