MBS-13930: Add Rockit to the Other Databases whitelist
This adds support for the Italian music industry database Rockit.

"Recording artist" pages (/$artistname) are allowed for artists, and
so are "performing artist" pages (/professionista/artistname).
Labels follow the same format as "recording artist" (/$labelname),
as do places (/$placename). A similar style is used for some festival entries
but it's unclear whether they should be events or series; this makes it
so that both are allowed, since the user has to decide on a case by case basis.
Finally, /$artistname/album/$albumname/$id is used for albums (RG level it seems).
I tried, but $albumname cannot be dropped even though there's an ID.

The site also includes reviews; we don't want to block them (we already
have a bunch) and they are easy to filter, so this also autoselects those.