MBS-1964: Allow opening "Add selected X for merging" in a new tab
This seems like the best option in order to allow editors to work
through the lists without having to keep going back to them.
Some editors wanted to keep the current option for some reason
so this adds a second button so that the editor has the choice.

In both cases, the checkboxes are cleared when submitting the merge,
allowing to easily continue using the page to merge further entries
if desired.

This was supposed to be achieved by the current returnToCurrentPage,
but that seems broken. And having two buttons gives users more of
a choice anyway; some browsers support shortcuts to open things
in new tabs, but not all of them do.