MBS-1964: Open "Add selected X for merging" in a new tab
This seems like the best option in order to allow editors to work
through the lists without having to keep going back to them.

This was supposed to be achieved by the current returnToCurrentPage,
but that seems broken. In any case, several users seem to prefer leaving
the previous page untouched (with checkboxes selected, etc) while
opening the merge form in a new tab.