MBS-13763: Add past year stats to the editors statistic page (#3465)
We've gotten requests for "top for last year" and it seems sensible;
one week is for big editing spurts only, and top of all time is basically
impossible for anyone who hasn't been editing for 15 years and/or
is kinda crazy in the coconut to get a good position on.

This adds top_yearly for both editing and voting. For editing, this counts
only applied edits (same as yearly and unlike weekly) since a year is
a long enough period that the open edits should not make that much
of a difference (and if they do, they won't a week later).