MBS-13914: Validate vote arguments for voter edit search (#3464)
This was causing ISEs when the passed argument was '' since
the query could crash with WHERE ().
But more generally, it seems any argument that is not a valid vote option
or "no" (which we use for searching for lack of votes) should just be deemed
invalid and reject the search, so this does that.