Fix incorrectly nested li elements
This was wrongly converted from the TT paginator, which just closes
the separator li before opening the next.
I'm not sure if this is even valid (probably not really?) but it doesn't seem
to have any purpose in any case and it triggers validation tests about the
two closing /li in a row.