Move IS[RW]C lists outside of SidebarProperties
SidebarProperties is a dl, and IsrcList / IswcList return a hydrated div
containing a separate dl. You cannot have the div inside a dl, but that's
not even the main problem, since you also cannot have nested dls...

This just needs to be alongside the SidebarProperties dl instead of
inside of it. Nothing seems to change in the display
on Chrome at least.