MBS-8668: Block adding duplicates to a series (JS)
Elsewhere, we accept relationships as different if their link order
varies. For series though, unless the number attribute is different,
we probably don't want to have the same entity listed many times
(and to the user is often not obvious at all that a link order even
exists, especially if the series is set to automated order).