Log errors in `get_chunked_with_retry`
I'm seeing these kinds of failures in the Selenium tests:

  [browser console log] [SEVERE]
  http://mbtest:5000/ws/js/artist/?q=gr%C3%B6up%20member&page=1&direct=false
  - Failed to load resource: the server responded with a status of 500
  (Internal Server Error)

The plackup logs show the request being dispatched, but the Solr logs don't
show any indication that it was received. I'm curious what LWP receives as the
response.

This logging should also be useful in production for debugging search errors.