Cache tests image without checkout
This partially reverts 81dec1dcc8976652981cb560f4c23a5228320964.

Since fb193a59f39fbb2e313f11be0e253bcf26c5c7d6, caching the tests image for
every commit (just so that it can be shared among the current jobs) is quite
expensive, storage-wise. GitHub Actions provides 10GB of cache for free, and
our tests images (with the MBS repository included) are close to 2GB each. And
that's not counting the Docker build cache, which is also stored on GHA.

It's not a proper use of the cache to begin with, since images cached by commit
hash are by definition transient, and can't be reused by other builds. (That's
what the Docker build cache I mentioned is for.)

This adds a hard-coded `TESTS_IMAGE_TAG` variable which is now used as the
cache key, and must be bumped by hand when the image changes. (Note that
`TESTS_IMAGE_TAG` must also be updated when Node modules change.) Of course,
the image is also no longer built with the current checkout baked in as before.

This has the effect of skipping the main `build-tests-image` steps if the image
is cached already, which saves time; on the flip side, each individual job now
has to perform a bit more setup work (creating the databases and compiling
static resources, for example), as those processes depend on the current
checkout, which must be copied into the image.