Fix `makes_no_changes` message
If you open e.g. an artist edit page and submit it with no changes, it should
reload and show a message to that effect. Instead it just directs you back to
the artist index.

I'm not sure how long that was broken, but the solution is to filter out
`undef` edits pushed by `edit_relationships`.