`Data::Relationship::load_subset`: Support loading by IDs
If you pass a HASH reference for `$target_types`, it will be assumed to be a
map of target types to relationship IDs. Only those relationships will be
loaded.

How does this compare to `Data::Relationship::get_by_ids`? With `load_subset`
you can load multiple relationships for multiple target types in one call. It
also sets the `source`, `target`, and `direction` attributes on the
relationships, and loads related info on them.