MBS-12170: Output containing country code for all artists in web service
This adds a `country` element/property to every artist in /ws/2 which returns
the ISO-3166-1 code of the country associated with the artist's area (i.e., the
country containing `artist.area`, or `artist.area` itself if it's a country).

Note that we did sometimes output a `country` element/property on artists prior
to this commit, but only in very particular circumstances: the artist's area
had to be loaded (which wasn't true for all artists in the response), and the
area itself had to be a country. This commit expands both the loading of this
information and the calculation of it (i.e., it now takes the area containment
into account).

The codes are loaded directly from the database and stored on each
`Entity::Artist` object. Note that `Data::Artist::load_country_code` also
caches the country codes in a `state` variable. This cache lasts as long as the
worker process does (refer to Starlet's min/max-reqs-per-child in production).