Alias `inc=artists` to `inc=artist-credits`
The difference between `inc=artists` and `inc=artist-credits` is very
ill-documented, inconsistent, and contradictory. It's better for everyone if we
just merge them so that they're aliases for each other.

I believe the original idea was that `artists` was for loading the top-level AC
(possibly with supplemental artist data?), and `artist-credits` was for loading
any sub-level, linked ACs. Except in a lot of places you'll see in this commit,
we were already checking `$inc->artists || $inc->artist_credits` anyway.

This will cause some additional data to be returned in a few cases where only
`inc=artists` was specified before.

Closes MBS-9005.