Split `_serialize_tags_and_ratings` in XMLSerializer.pm
In a subsequent commit, artist credits for subqueried entities will be loaded
in batch with other artist credits in the response. That may cause tags and
ratings to be outputted in places where they previously weren't, because the
code simply outputs them if they're loaded at all in those cases, not depending
on the context.

So splitting these is needed so that we can make `_serialize_tags_and_genres`
conditional in `_serialize_artist` for subqueried entities (release group
releases and recording releases) without affecting ratings, which don't have
the same restrictions for some reason.