MBS-13969: Preserve isrcs and iswcs when missing from POST request
Turns the `isrcs` and `iswcs` fields into repeatable compound fields which
require a `removed` subfield to be set in order to remove them, e.g.,

  isrcs.0=ZZ00Z0000001
  isrcs.0.removed=1

Notice that we still explicitly support seeding and submitting `isrcs.N` and
`iswcs.N` as repeatable text fields. Even the edit forms still submit them like
that. `HTML::FormHandler` /really/ doesn't like that. We need some hacks to
support it. Internally, `$identifier.N` is converted to `$identifier.N.value`
before validation.

Removing an identifier from the editing page injects the relevant hidden fields
in its place if that identifier was part of the "current" set of identifiers
stored in the database when the form was loaded. No special processing is done
on these hidden fields prior to submission to check whether a removed field was
added again. That's handled by the server. (`removed` is explicitly ignored if
the same identifier is being added by another index.)