MBS-13711 (I): Warn about improperly-capitalized ETI (#3504)
Make the tracklist editor display a warning if a track title
appears to contain ETI with an incorrectly-capitalized
non-title word, e.g. "Remix" or "Version". If the user is a
beginner, a checkbox is displayed for them to confirm that
the usage is correct.

This check is skipped for releases that are set to a
language other than English and for Japan-only releases.