Use row ids rather than passing reltype gid
mwiencek suggested using the rowids we already have rather than
passing the gids as well for the check. As such, this adds a new
PART_OF_SERIES_LINK_TYPE_IDS constant; PART_OF_SERIES
is unused now on the Perl side, but it already was before I started
working on this patch so that might be fine.