MBS-13964: Some recordings are missing a first release date
When releases are merged with the "append" strategy, then mediums are moved by
updating their `release` column. (See the `Data::Release::merge` method.)

Since this affects the releases a recording appears on, we need a trigger on
the `medium` table to update `recording_first_release_date` whenever a
`medium.release` column is updated.

The upgrade script also takes care of recalculating all existing data in the
materialized table, provided it's non-empty to begin with.