Never override `timeout` on `$context->lwp`
The server context's `lwp` object persists in the current worker process
by design. Overriding the timeout (and not resetting it back to the
default) introduces random behavior that depends on the order of
requests; in one case this has very surprising results:

In `ExternalUtils::get_chunked_with_retry`, the following code
calculates the number of request attempts remaining based on the LWP
timeout:

  my $retries_remaining = int(25.0 / $ua->timeout);

Note that if the timeout is 30, as it is set by the artwork uploading
code, `$retries_remaining` will be initialized to 0, and no requests
will be performed. This is in fact the source of semi-frequent transient
search failures in the Selenium tests, and I believe the same issue can
be reproduced in production too.

That's not to say the `$retries_remaining` calculation makes sense to
begin with; I'll be changing it to a fixed number in a subsequent commit
to avoid such issues in the future.